[MOD/#293] - Task: command / query service 구분
